Fixed some crash bugs for autofill
