+++ evervault-inputs/src/main/java/com/evervault/sdk/input/ui/component/AutofillComponent.kt
@@ -1,0 +1,32 @@
+package com.evervault.sdk.input.ui.component
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.autofill.AutofillNode
+import androidx.compose.ui.autofill.AutofillType
+import androidx.compose.ui.layout.boundsInWindow
+import androidx.compose.ui.layout.onGloballyPositioned
+import androidx.compose.ui.platform.LocalAutofillTree
+
+
+@ExperimentalComposeUiApi
+@Composable
+internal fun AutofillComponent(
+    modifier: Modifier,
+    autofillTypes: List<AutofillType>,
+    onFill: ((String) -> Unit),
+    content: @Composable (AutofillNode) -> Unit
+) {
+    val autofillNode = AutofillNode(onFill = onFill, autofillTypes = autofillTypes)
+    LocalAutofillTree.current.plusAssign(autofillNode)
+
+    Box(
+        modifier.onGloballyPositioned {
+            autofillNode.boundingBox = it.boundsInWindow()
+        }
+    ) {
+        content(autofillNode)
+    }
+}
