+++ evervault-inputs/src/main/java/com/evervault/sdk/input/ui/component/CustomTextField.kt
@@ -15,6 +15,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.autofill.AutofillType
+import androidx.compose.ui.platform.LocalAutofill
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardCapitalization
@@ -22,7 +23,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
-import com.evervault.sdk.input.ui.modifier.autofill
+import com.evervault.sdk.input.ui.modifier.autofillOnFocusChange
 
 /**
  * Input field that accepts composables for the placeholder and labels, so the user can modify and provide them as they need.
@@ -40,38 +41,42 @@ internal fun CustomTextField(
     autofillType: AutofillType
 ) {
     val interactionSource = remember { MutableInteractionSource() }
+    val autofill = LocalAutofill.current
 
-    BasicTextField(
-        value = state.value,
-        onValueChange = { state.value = it },
-        modifier = modifier.autofill(
-            autofillTypes = listOf(autofillType),
-            onFill = { state.value = TextFieldValue(it) }
-        ),
-        textStyle = textStyle,
-        interactionSource = interactionSource,
-        keyboardOptions = KeyboardOptions(
-            capitalization = KeyboardCapitalization.None,
-            autoCorrect = false,
-            keyboardType = KeyboardType.Number,
-            imeAction = if (onNext == null) ImeAction.Done else ImeAction.Next
-        ),
-        keyboardActions = KeyboardActions(
-            onNext = { onNext?.invoke() },
-        ),
-        decorationBox = @Composable { innerTextField ->
-            TextFieldDefaults.DecorationBox(
-                value = state.value.text,
-                innerTextField = innerTextField,
-                enabled = true,
-                singleLine = true,
-                visualTransformation = VisualTransformation.None,
-                interactionSource = interactionSource,
-                label = label,
-                placeholder = placeholder,
-                colors = textFieldColors,
-                contentPadding = PaddingValues(0.dp),
-            )
-        }
-    )
+    AutofillComponent(
+        modifier = modifier,
+        autofillTypes = listOf(autofillType),
+        onFill = { state.value = TextFieldValue(it) }
+    ) { autofillNode ->
+        BasicTextField(
+            value = state.value,
+            onValueChange = { state.value = it },
+            modifier = modifier.autofillOnFocusChange(autofill, autofillNode),
+            textStyle = textStyle,
+            interactionSource = interactionSource,
+            keyboardOptions = KeyboardOptions(
+                capitalization = KeyboardCapitalization.None,
+                autoCorrect = false,
+                keyboardType = KeyboardType.Number,
+                imeAction = if (onNext == null) ImeAction.Done else ImeAction.Next
+            ),
+            keyboardActions = KeyboardActions(
+                onNext = { onNext?.invoke() },
+            ),
+            decorationBox = @Composable { innerTextField ->
+                TextFieldDefaults.DecorationBox(
+                    value = state.value.text,
+                    innerTextField = innerTextField,
+                    enabled = true,
+                    singleLine = true,
+                    visualTransformation = VisualTransformation.None,
+                    interactionSource = interactionSource,
+                    label = label,
+                    placeholder = placeholder,
+                    colors = textFieldColors,
+                    contentPadding = PaddingValues(0.dp),
+                )
+            }
+        )
+    }
 }
